requirements.txt
aws-cdk-lib

aws_cdk.core
 
aws_cdk.aws_s3
aws_cdk.aws_lambda